zatsuyo/requirements.txt
discord.py[voice]
requests==2.12.4
pip
opuslib